Add support for trace version negotiation
We send ASCII 'VERS' as the insn field in the initial reset field to indicate
that this is a version negotiation packet. We then use the high bits of
the initial packet's rvfi_halt field as an indicator of the version:
(halt >> 1) + 1, so unchanged implementations will report version 1
support whereas newer implementations can send a value of '3' to indicate
that they support version 2.