Fix cache templates and add CHERI-64 cache testing
Previously, the addresses used in the cache template were quite likely
to be unaligned and also out of the RVFI-DII memory region.
This removes the unaligned-ness (apart from in CHERI-128) and removes
the out-of-region accesses.
This also updates the template to use "store double" rather than "store
quad" for CHERI-64.